requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
mysql-connector-python==8.0.28
    # via -r requirements.in
protobuf==3.20.1
    # via mysql-connector-python
